+++ yourdio.py
@@ -17,7 +17,7 @@ Author: Yourdio Contributors
 License: MIT
 """
 
-__version__ = "1.0.4"
+__version__ = "1.0.5"
 __author__ = "Yourdio Contributors"
 __license__ = "MIT"